backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.4
cssselect==1.2.0
datasets==3.5.0
dateparser==1.2.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
fastapi-mail==1.5.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.44.0
lxml==4.9.3
lxml_html_clean==0.4.2
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.34.1
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.2
numpy==2.2.4
openai==1.3.0
openai-whisper==20240930
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pygooglenews==0.1.3
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.33.0
sentencepiece==0.2.0
sentry-sdk==2.25.1
seqeval==1.2.2
setproctitle==1.3.5
setuptools==78.1.0
sgmllib3k==1.0.0
simpletransformers==0.70.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.44.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
wandb==0.19.9
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
